requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
nicegui
httpx
pytz
requests
python-jose[cryptography]
passlib[bcrypt]
python-multipart